fix: ensure that default approle exists and is unique on updates
If migrating an existing application, it may define an AppRole that uses
the same `value` (i.e. the string-value emitted in the `roles` claim),
which will cause collisions with our hardcoded values for the default
AppRole.

This ensures that such duplicate AppRoles are disabled and subsequently
removed from the application, replaced by the hardcoded default AppRole
that should be consistent across all Applications instead.