feat: require approle assignment when groups assignment feature is enabled
This enables Azure AD to enforce authorization on service-to-service
token requests (client credentials flow), as well as enabling users to
restrict logins to explicitly assigned groups to the application.

If no groups are defined in the custom resource, or all the groups
provided do not exist, we will fall back to assigning a group that
should contain _all_ the users in the tenant. This emulates the
previous behaviour where all users in the tenant can log in through
the application.