fix: ensure oauth2permissionscopes and approle IDs are valid
Microsoft requires us to disable both scopes and roles that are to be removed.
As we've set some default scopes that replaces any existing scopes, we must
ensure that all other scopes (if present) must be disabled.
As for the AppRoles, we reuse any existing role that has the same emitted
Value as our default role if found. Otherwise, append the default role.

This fix will primarily target and enable migration from our legacy
provisioning systems.