fix(synchronizer): allow multi-value resync annotations for concurrent events
In cases where an event triggers a resync, we mark a matching resource
by applying annotation. The annotation was previously removed in its entirety
after processing. If multiple events occur (almost) simultaneously and result in
marking the same resource for resync, this could result in only one of
the events being processed.

This commit fixes this by appropriately appending and removing the
annotation values from a list of multiple values. This should remediate the
case where a resync annotation is applied to a resoruce while the annotation
already has a value, i.e. it is currently being processed - which
effectively should process all the events that each should've triggered
a resync.