feat(custom-claims-policy): remove per-resource option for policy assignment
This is handled globally by the controller through an option.

The spec gives an illusion of choice as it accepts a list of claims.
However, a service principal can only have one policy assigned at any
time. Configuring multiple claims means that we have to have one unique
policy per unique combination of such claims. We could replace this with
a boolean toggle per resource, but maintaining this centrally through
the operator enforces a proper default - which hopefully leads to less
confusion for the end-user developers.