fix(azure/client): set acceptMappedClaims after creation
Also only set if the claims mapping policy feature is enabled.

Setting this field during creation of an Application
results in an a 403 error if the calling principal
only has the Application.ReadWrite.OwnedBy permission.

We are however, allowed to _update_ an Application with the
same field set to true after its creation.