fix(azure/client): assign ownership to authenticated principal
Assign the authenticated principal used by the operator
as an owner of the managed applications and service principals,
if not already present.

We do this to allow usage of the `Application.ReadWrite.OwnedBy`
permission for the authenticated principal.
This will not work if the principal is not an owner of the
resources, so you may need to temporarily add `Application.ReadWrite.All`
and resynchronize all applications to ensure ownership is assigned.

Fixes https://github.com/nais/azurerator/issues/164.