feat: resync matching applications in same cluster on create
This modifies the reconciler to produce an event on application creation
to the synchronizer for local (e.g. same cluster) resyncs.

The Kafka variant is required when synchronizing such events across multiple
clusters.

A caveat with this in-process approach is that the event is lost should
the process restart or shut down before a resync is successfully
triggered, though we expect this to only happen in very rare
circumstances (e.g. the Kubernetes API server is unavailable or
unreachable)